+++ bigml/api.py
@@ -57,7 +57,7 @@ from bigml.batchanomalyscorehandler import BatchAnomalyScoreHandler
 from bigml.projecthandler import ProjectHandler
 from bigml.samplehandler import SampleHandler
 from bigml.correlationhandler import CorrelationHandler
-from bigml.testhandler import TestHandler
+from bigml.statisticaltesthandler import StatisticalTestHandler
 
 # Repeating constants and functions for backwards compatibility
 
@@ -79,7 +79,7 @@ from bigml.resourcehandler import (
     BATCH_PREDICTION_PATH, CLUSTER_PATH, CENTROID_PATH, BATCH_CENTROID_PATH,
     ANOMALY_PATH, ANOMALY_SCORE_PATH, BATCH_ANOMALY_SCORE_PATH, PROJECT_PATH,
     SAMPLE_PATH, SAMPLE_RE, CORRELATION_PATH, CORRELATION_RE,
-    TEST_PATH, TEST_RE)
+    STATISTICAL_TEST_PATH, STATISTICAL_TEST_RE)
 
 
 from bigml.resourcehandler import (
@@ -89,7 +89,7 @@ from bigml.resourcehandler import (
     get_prediction_id, get_batch_prediction_id, get_batch_centroid_id,
     get_batch_anomaly_score_id, get_resource_id, resource_is_ready,
     get_status, check_resource, http_ok, get_project_id, get_sample_id,
-    get_correlation_id, get_test_id)
+    get_correlation_id, get_statistical_test_id)
 
 # Map status codes to labels
 STATUSES = {
@@ -114,7 +114,7 @@ def count(listing):
         return listing['meta']['query_total']
 
 
-class BigML(TestHandler, CorrelationHandler, SampleHandler, ProjectHandler,
+class BigML(StatisticalTestHandler, CorrelationHandler, SampleHandler, ProjectHandler,
             BatchAnomalyScoreHandler, BatchCentroidHandler,
             BatchPredictionHandler, EvaluationHandler, AnomalyScoreHandler,
             AnomalyHandler, CentroidHandler, ClusterHandler, PredictionHandler,
@@ -177,7 +177,7 @@ class BigML(TestHandler, CorrelationHandler, SampleHandler, ProjectHandler,
         ProjectHandler.__init__(self)
         SampleHandler.__init__(self)
         CorrelationHandler.__init__(self)
-        TestHandler.__init__(self)
+        StatisticalTestHandler.__init__(self)
 
         self.getters = {}
         for resource_type in RESOURCE_RE:
@@ -233,8 +233,8 @@ class BigML(TestHandler, CorrelationHandler, SampleHandler, ProjectHandler,
                     return resource['object']['clusters']['fields']
                 elif CORRELATION_RE.match(resource_id):
                     return resource['object']['correlations']['fields']
-                elif TEST_RE.match(resource_id):
-                    return resource['object']['tests']['fields']
+                elif STATISTICAL_TEST_RE.match(resource_id):
+                    return resource['object']['statistical_tests']['fields']
                 elif SAMPLE_RE.match(resource_id):
                     return dict([(field['id'], field) for field in
                                  resource['object']['sample']['fields']])