+++ bigml/resourcehandler.py
@@ -44,7 +44,7 @@ BATCH_ANOMALY_SCORE_PATH = 'batchanomalyscore'
 PROJECT_PATH = 'project'
 SAMPLE_PATH = 'sample'
 CORRELATION_PATH = 'correlation'
-TEST_PATH = 'test'
+STATISTICAL_TEST_PATH = 'statisticaltest'
 
 # Resource Ids patterns
 ID_PATTERN = '[a-f0-9]{24}'
@@ -72,7 +72,8 @@ BATCH_ANOMALY_SCORE_RE = re.compile(r'^%s/%s$' % (BATCH_ANOMALY_SCORE_PATH,
 PROJECT_RE = re.compile(r'^%s/%s$' % (PROJECT_PATH, ID_PATTERN))
 SAMPLE_RE = re.compile(r'^%s/%s$' % (SAMPLE_PATH, ID_PATTERN))
 CORRELATION_RE = re.compile(r'^%s/%s$' % (CORRELATION_PATH, ID_PATTERN))
-TEST_RE = re.compile(r'^%s/%s$' % (TEST_PATH, ID_PATTERN))
+STATISTICAL_TEST_RE = re.compile(r'^%s/%s$' % \
+    (STATISTICAL_TEST_PATH, ID_PATTERN))
 
 RESOURCE_RE = {
     'source': SOURCE_RE,
@@ -91,13 +92,15 @@ RESOURCE_RE = {
     'project': PROJECT_RE,
     'sample': SAMPLE_RE,
     'correlation': CORRELATION_RE,
-    'test': TEST_RE}
+    'statisticaltest': STATISTICAL_TEST_RE}
 
 RENAMED_RESOURCES = {
     'batchprediction': 'batch_prediction',
     'batchcentroid': 'batch_centroid',
     'anomalyscore': 'anomaly_score',
-    'batchanomalyscore': 'batch_anomaly_score'}
+    'batchanomalyscore': 'batch_anomaly_score',
+    'statisticaltest': 'statistical_test'
+}
 
 NO_QS = [EVALUATION_RE, PREDICTION_RE, BATCH_PREDICTION_RE,
          CENTROID_RE, BATCH_CENTROID_RE, ANOMALY_SCORE_RE,
@@ -296,11 +299,11 @@ def get_correlation_id(correlation):
     return get_resource(CORRELATION_RE, correlation)
 
 
-def get_test_id(test):
-    """Returns a test/id.
+def get_statistical_test_id(statistical_test):
+    """Returns a statisticaltest/id.
 
     """
-    return get_resource(TEST_RE, test)
+    return get_resource(STATISTICAL_TEST_RE, statistical_test)
 
 
 def get_resource_id(resource):