+++ bigml/tests/world.py
@@ -45,18 +45,20 @@ RESOURCE_TYPES = [
     'project',
     'sample',
     'correlation',
-    'test']
+    'statisticaltest']
 IRREGULAR_PLURALS = {
     'anomaly': 'anomalies',
     'batchprediction': 'batch_predictions',
     'batchcentroid': 'batch_centroids',
     'anomalyscore': 'anomaly_scores',
-    'batchanomalyscore': 'batch_anomaly_scores'}
+    'batchanomalyscore': 'batch_anomaly_scores',
+    'statisticaltest': 'statistical_tests'}
 TRANSLATED_RESOURCES = {
     'batchprediction': 'batch_prediction',
     'batchcentroid': 'batch_centroid',
     'anomalyscore': 'anomaly_score',
-    'batchanomalyscore': 'batch_anomaly_score'}
+    'batchanomalyscore': 'batch_anomaly_score',
+    'statisticaltest': 'statistical_test'}
 
 
 def plural(resource_type):